backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.0
pypdf==4.2.0
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.3.0
google-genai>=1.0.0
python-dotenv==1.0.1
pandas==2.2.2
aiofiles==23.2.1
numpy>=1.26.0

